se cambio formato de precio a precio regional y se modifico interfaz de detalle producto

+++ app/src/main/java/com/example/milsaborescompose/data/local/Product.kt
@@ -8,7 +8,7 @@ data class Product(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val name: String,
     val description: String,
-    val price: Double,
+    val price: Int,
     val image: String,
     val category: String
 )
